DFE-524 - moving the editable parts of a release into their own component
